Summary:

Test Plan:
python test/test_quantization.py TestQuantizeScriptPTSQOps.test_quantized_conv1d

Reviewers:

Subscribers:

Tasks:

Tags:

[ghstack-poisoned]

+++ aten/src/ATen/native/quantized/cpu/qnnpack_utils.h
@@ -88,7 +88,7 @@ struct PackedConvWeightsQnnp : public ConvPackedParamsBase<kSpatialDim> {
       torch::List<int64_t> padding,
       torch::List<int64_t> dilation,
       int64_t groups,
-      c10::optional<float> input_scale,
+      c10::optional<double> input_scale,
       std::vector<int64_t> kernel,
       float w_scale,
       int32_t w_zp)
@@ -111,7 +111,7 @@ struct PackedConvWeightsQnnp : public ConvPackedParamsBase<kSpatialDim> {
   torch::List<int64_t> padding_;
   torch::List<int64_t> dilation_;
   int64_t groups_;
-  c10::optional<float> input_scale;
+  c10::optional<double> input_scale;
   std::vector<int64_t> kernel;
   float w_scale;
   int32_t w_zp;
